Adjusting frontend services to apply the changes in backend functions. Moved system config template in a specific folder
