+++ frontend/services/onboarding/advanceOnboardingStatus.js
@@ -4,10 +4,10 @@ import { ONBOARDING_ERROR_MESSAGES } from '@/constants/onboarding';
 
 import { functions } from '@/redux/store';
 
-const advanceOnboardingStatus = async (uid) => {
+const advanceOnboardingStatus = async (uid, step) => {
   try {
     const advanceStatus = httpsCallable(functions, 'advanceOnboardingStatus');
-    const response = await advanceStatus({ uid });
+    const response = await advanceStatus({ uid, step });
 
     return response.data;
   } catch (error) {
